requirements.txt
kafka-python
pandas
requests
apache-beam[gcp]
google-cloud
spark-bigquery-connector